+++ src/main/java/com/phone_shop/phoneshop/dto/ModelDTO.java
@@ -4,8 +4,8 @@ import lombok.Data;
 
 @Data
 public class ModelDTO {
-    private Integer id;
+    private Long id;
     private String name;
-    private Integer brandId;
+    private Long brandId;
 
 }
