+++ src/main/java/com/phone_shop/phoneshop/repository/ModelRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ModelRepository extends JpaRepository<Model, Integer> {
-    List<Model> findByBrandId(Integer id);
+public interface ModelRepository extends JpaRepository<Model, Long> {
+    List<Model> findByBrandId(Long id);
 }
