+++ src/main/java/com/phone_shop/phoneshop/service/util/ResponseHelper.java
@@ -5,7 +5,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 
 public class ResponseHelper {
-    public static ResponseEntity<?> deleteSuccess(String entity, Integer id) {
+    public static ResponseEntity<?> deleteSuccess(String entity, Long id) {
         ResourceData resourceData = new ResourceData(
                 HttpStatus.ACCEPTED, true, "%s with id =%d delete Success".formatted(entity, id)
